refatoração da service de métricas
